:sparkles: Add very faded color
We needed to add a more faded color to use in the ground `NonText`.
Previously, we are using the faded color, which is the same as the
string. Hence, when using inlay hints, the output was very confusing.

Those colors were design by increasing / decrasing the color luminosity
in the light / dark themes.